perf(api/): :fire: delete deprecated code
