+++ api/src/controllers/login.controller.js
@@ -1,21 +1,9 @@
-const {
-  findOneStudentByProfile,
-  findOneProfessionalByProfile,
-  findOneCompany,
-  findUser,
-  findAccount
-} = require("../helpers/findUser");
+const { findAccount } = require("../helpers/findAccount");
 const { createToken } = require("../helpers/jwt");
-
-const userType = (type) => {
-  if (type === "company") return findOneCompany;
-  return findUser;
-};
 
 const validateUser = async (user) => {
   const { email, password } = user;
   try {
-    // const findAccount = userType(type);
     const foundedAccount = await findAccount(email);
     // console.log(foundedAccount);
     if (!foundedAccount) return;
